requirements.txt
# Mahabharat Shorts Pipeline — Python Dependencies
# Install with: pip install -r requirements.txt

# ── Core ────────────────────────────────────────
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0

# ── YouTube Upload ──────────────────────────────
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# ── Utilities ───────────────────────────────────
pathlib2>=2.3.7
